rename art components
