+++ src/components/art/BubblesArt.tsx
