+++ src/components/art/StrataArt.tsx
