+++ src/components/art/SupershapeArt.tsx
