+++ src/components/art/WaveArt.tsx
