+++ src/components/art/styleRegistry.tsx
@@ -1,9 +1,9 @@
-import OrbitArt from "./Orbits";
-import StrataArt from "./Strata";
-import ConstellationArt from "./Constellation";
-import BubblesArt from "./Bubbles";
-import WaveArt from "./Wave";
-import SupershapeArt from "./Supershape";
+import OrbitArt from "./OrbitArt";
+import StrataArt from "./StrataArt";
+import ConstellationArt from "./ConstellationArt";
+import BubblesArt from "./BubblesArt";
+import WaveArt from "./WaveArt";
+import SupershapeArt from "./SupershapeArt";
 import type { ArtStyleProps } from "../../logic/types";
 
 export const STYLES: Record<
